moved shared content key
